fix format and warnings

+++ JavaFunctionalCpp/lexer.hpp
@@ -13,7 +13,7 @@ private:
 	char peekNext();
 	char lookAhead(int offset);
 	void advance();
-	size_t index = 0;
+	int index = 0;
 	std::string program;
 
 	unsigned int row = 0;
